GUI combine Aiden's work on save/load
